yaw and zoom now working
