--- f360-script/.env
@@ -1,1 +1,0 @@
-PYTHONPATH=C:/Users/gena0/AppData/Local/Autodesk/webdeploy/production/525fe9b2ece26c4d179982b11ecdfb713cb5058f/Api/Python/packages
