--- f360-script/External controls.manifest
@@ -1,13 +1,0 @@
-{
-	"autodeskProduct":	"Fusion360",
-	"type":	"addin",
-	"id":	"0afdf698-8eed-4eaa-ac5c-d6ae17db8147",
-	"author":	"InTostor",
-	"description":	{
-		"":	""
-	},
-	"version":	"0.1",
-	"runOnStartup":	true,
-	"supportedOS":	"windows|mac",
-	"editEnabled":	true
-}
