--- f360-script/lib/fusion360utils/__init__.py
@@ -1,2 +1,0 @@
-from .general_utils import *
-from .event_utils import *
