--- f360-script/lib/logger.py
@@ -1,28 +1,0 @@
-import adsk.core
-
-class UiLogger:
-    def __init__(self, forceUpdate):  
-        app = adsk.core.Application.get()
-        ui  = app.userInterface
-        palettes = ui.palettes
-        self.textPalette = palettes.itemById("TextCommands")
-        self.forceUpdate = forceUpdate
-        self.textPalette.isVisible = True 
-    
-    def print(self, text):       
-        self.textPalette.writeText(text)
-        if (self.forceUpdate):
-            adsk.doEvents() 
-
-class FileLogger:
-    def __init__(self, filePath): 
-        try:
-            open(filePath, 'a').close()
-        
-            self.filePath = filePath
-        except:
-            raise Exception("Could not open/create file = " + filePath)
-
-    def print(self, text):
-        with open(self.filePath, 'a') as txtFile:
-            txtFile.writelines(text + '\r\n')
